scripts/caffe/caffe_py27_requirements.txt
absl-py==0.1.13
astor==0.6.2
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
cycler==0.10.0
Cython==0.28.4
dask==0.18.2
decorator==4.3.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
graphviz==0.8.4
grpcio==1.10.1
h5py==2.8.0
html5lib==0.9999999
idna==2.6
imutils==0.4.6
ipython==5.7.0
ipython-genutils==0.2.0
Keras==2.1.5
kiwisolver==1.0.1
leveldb==0.194
lmdb==0.94
Markdown==2.6.11
matplotlib==2.2.2
mock==2.0.0
networkx==2.1
nose==1.3.7
numpy==1.15.0
olefile==0.44
pandas==0.23.3
pathlib2==2.3.2
pbr==4.2.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
progressbar2==3.37.0
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
pydot==1.2.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.5.0
python-gflags==3.1.2
python-utils==2.3.0
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.13
requests==2.18.4
scandir==1.7
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
stevedore==1.29.0
subprocess32==3.5.2
tensorboard==1.7.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
toolz==0.9.0
traitlets==4.3.2
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wcwidth==0.1.7
Werkzeug==0.14.1
